refactor: print 및 input 관련 로직을 view 로 이동
