+++ src/domain/CarRace.js
@@ -1,8 +1,6 @@
 import { ERROR_CAR_RACE_COUNT_NOT_VALID } from "../constants/error.js";
-import { readLineAsync } from "../utils/io.js";
-import { joinCarNamesByComma, printCarsStatus } from "../utils/cars.js";
 import { isNumber } from "../utils/number.js";
-
+import { printCarsStatus } from "../utils/view.js";
 class CarRace {
   #remainingRaceCount;
 
@@ -61,27 +59,6 @@ class CarRace {
       this.#remainingRaceCount--;
     }
   }
-
-  static async getTotalRaceCountUntilValid() {
-    while (true) {
-      const input = await readLineAsync("시도할 회수는 몇회인가요?");
-      const totalCount = Number(input);
-
-      try {
-        CarRace.validateTotalRaceCount(totalCount);
-        return totalCount;
-      } catch (e) {
-        console.log(ERROR_CAR_RACE_COUNT_NOT_VALID);
-        console.log();
-      }
-    }
-  }
-
-  printWinners = () => {
-    const joinedCarRaceWinners = joinCarNamesByComma(this.winners);
-
-    console.log(`${joinedCarRaceWinners}가 최종 우승했습니다.`);
-  };
 }
 
 export default CarRace;
